Add BaseReticleWorker and inherit
